Implemented suggested changes and moved the folder into separate directory
