+++ app/exerlog/lib/UI/calendar/view/calendar_page.dart
@@ -1,0 +1,47 @@
+import 'package:exerlog/UI/calendar/widgets/calendar_widget.dart';
+import 'package:exerlog/UI/calendar/widgets/logout_button.dart';
+import 'package:exerlog/UI/global.dart';
+import 'package:exerlog/UI/gradient_button.dart';
+import 'package:exerlog/UI/workout/workout_page.dart';
+import 'package:flutter/material.dart';
+
+class CalendarPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: backgroundColor,
+      appBar: AppBar(
+        title: Text("Exerlog"),
+        backgroundColor: backgroundColor.withOpacity(0.75),
+        actions: [
+          LogoutButton(),
+        ],
+      ),
+      body: Padding(
+        padding: EdgeInsets.all(20),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            CalendarWidget(),
+            Container(
+              width: MediaQuery.of(context).size.width * 0.8,
+              child: RaisedGradientButton(
+                child: Text(
+                  "START NEW WORKOUT",
+                  style: buttonTextSmall,
+                ),
+                onPressed: () {
+                  Navigator.of(context).pushReplacement(
+                    MaterialPageRoute(
+                      builder: (context) => WorkoutPage(null),
+                    ),
+                  );
+                },
+              ),
+            )
+          ],
+        ),
+      ),
+    );
+  }
+}
