+++ app/exerlog/lib/UI/gradient_button.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 class RaisedGradientButton extends StatefulWidget {
