+++ app/exerlog/lib/main.dart
@@ -1,28 +1,19 @@
 // @dart = 2.9
 import 'package:exerlog/Bloc/authentication.dart';
-import 'package:exerlog/Bloc/user_bloc.dart';
-import 'package:exerlog/Bloc/workout_bloc.dart';
-import 'package:exerlog/Models/exercise.dart';
-import 'package:exerlog/Models/sets.dart';
-import 'package:exerlog/Models/workout.dart';
-import 'package:exerlog/UI/calendar/calendar_page.dart';
-import 'package:exerlog/UI/workout/workout_page.dart';
+import 'package:exerlog/UI/calendar/view/calendar_page.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'UI/global.dart';
-// import '/UI/login_page.dart';
-import 'Models/user.dart';
-import 'UI/login_screen/login_page.dart';
 import 'package:flutter/services.dart';
+import 'UI/login_screen/login_page.dart';
 
 String userID = '';
 User the_user;
+
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
-  SystemChrome.setPreferredOrientations(
-    [DeviceOrientation.portraitUp,]
-  );
+  SystemChrome.setPreferredOrientations([
+    DeviceOrientation.portraitUp,
+  ]);
   runApp(MyApp());
 }
 
@@ -44,23 +35,23 @@ class MyApp extends StatelessWidget {
           User user = FirebaseAuth.instance.currentUser;
           if (user != null) {
             return MaterialApp(
-            debugShowCheckedModeBanner: false,
-            theme: ThemeData(
-              primarySwatch: Colors.blue,
-            ),
-            home: CalendarPage(),
-            color: Colors.blue,
-          );
+              debugShowCheckedModeBanner: false,
+              theme: ThemeData(
+                primarySwatch: Colors.blue,
+              ),
+              home: CalendarPage(),
+              color: Colors.blue,
+            );
           } else {
-          return MaterialApp(
-            debugShowCheckedModeBanner: false,
-            theme: ThemeData(
-              primarySwatch: Colors.blue,
-            ),
-            home: LoginPage('1'),
-            color: Colors.blue,
-          );
-        }
+            return MaterialApp(
+              debugShowCheckedModeBanner: false,
+              theme: ThemeData(
+                primarySwatch: Colors.blue,
+              ),
+              home: LoginPage('1'),
+              color: Colors.blue,
+            );
+          }
         }
 
         // Otherwise, show something whilst waiting for initialization to complete
